src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv ==1.0.1
pydantic-settings==2.5.2
aiofiles==24.1.0
langchain==0.3.0
PyMuPDF==1.24.10


